refactor: move internal/janitor to internal/worker/janitor
